requirements.txt
python-dotenv
supabase
requests
google-genai
python-slugify
beautifulsoup4
elevenlabs
fastapi
uvicorn[standard]
jinja2
python-multipart
